Queries UtilityAPI every hour and populates DB with new readings
